Add common dismissNotification function for `onNewIntent()` and `onActivityCreated()`
