+++ sample/src/main/java/com/clevertap/demo/MyApplication.kt
@@ -5,7 +5,6 @@ import android.app.Application.ActivityLifecycleCallbacks
 import android.app.NotificationManager
 import android.content.Context
 import android.content.Intent
-import android.os.StrictMode
 import android.os.Bundle
 import android.util.Log
 import androidx.multidex.MultiDex
@@ -18,9 +17,7 @@ import com.clevertap.android.sdk.CleverTapAPI.LogLevel.VERBOSE
 import com.clevertap.android.sdk.SyncListener
 import com.clevertap.android.sdk.interfaces.NotificationHandler
 import com.clevertap.android.sdk.pushnotification.CTPushNotificationListener
-import com.clevertap.android.sdk.pushnotification.PushConstants.ALL_DEVICES
-import com.clevertap.android.sdk.pushnotification.PushConstants.NO_DEVICES
-import com.clevertap.android.sdk.pushnotification.PushConstants.XIAOMI_MIUI_DEVICES
+import com.clevertap.demo.ui.main.NotificationUtils
 import com.google.android.gms.security.ProviderInstaller
 import com.google.android.gms.security.ProviderInstaller.ProviderInstallListener
 import org.json.JSONObject
@@ -110,6 +107,7 @@ class MyApplication : MultiDexApplication(), CTPushNotificationListener, Activit
     }
 
     override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {
+        NotificationUtils.dismissNotification(activity.intent,applicationContext)
     }
 
     override fun onActivityStarted(activity: Activity) {
